requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
sse-starlette